+++ tests/common.py
@@ -4,6 +4,9 @@ import pytest
 from conftest import TestData
 from conftest import under_uwsgi
 
+from cachelib import MemcachedCache
+from cachelib import SimpleCache
+
 
 class CommonTests(TestData):
     """A base set of tests to be run for all cache types"""
@@ -75,3 +78,36 @@ class CommonTests(TestData):
         for k, v in self.sample_pairs.items():
             assert cache.get(f"{k}-t0") == v
             assert not cache.get(f"{k}-t1")
+
+    def test_signed_set_get(self):
+        if isinstance(self.cache_factory(), (MemcachedCache, SimpleCache)):
+            pytest.skip("Simple and MemcachedCache do not support signing.")
+        cache = self.cache_factory(secret_key="not very secret")
+
+        for k, v in self.sample_pairs.items():
+            assert cache.set(k, v)
+            assert cache.get(k) == v
+
+    def test_signed_does_not_get_unsigned(self):
+        unsigned_cache = self.cache_factory()
+        if isinstance(unsigned_cache, (MemcachedCache, SimpleCache)):
+            pytest.skip("Simple and MemcachedCache do not support signing.")
+
+        signed_cache = self.cache_factory(secret_key="not very secret")
+
+        for k, v in self.sample_pairs.items():
+            assert unsigned_cache.set(k, v)
+            assert signed_cache.get(k) is None
+
+    def test_signed_does_not_get_wrong_key(self):
+        if isinstance(self.cache_factory(), (MemcachedCache, SimpleCache)):
+            pytest.skip("Simple and MemcachedCache do not support signing.")
+
+        signed_cache1 = self.cache_factory(secret_key="not very secret")
+        signed_cache2 = self.cache_factory(secret_key="another not secret value")
+
+        for k, v in self.sample_pairs.items():
+            assert signed_cache1.set(k, v)
+            assert signed_cache2.get(k) is None
+            assert signed_cache2.set(k, v)
+            assert signed_cache1.get(k) is None